% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbs_norm.R
\name{gbs_norm}
\alias{gbs_norm}
\title{Gibbs Algorithm for Generating Multivariate Normal Distribution}
\usage{
gbs_norm(n, mu, A, init, burn = 1000)
}
\arguments{
\item{n}{The numbers of samples one wants to obtain.}

\item{mu, A}{Parameters for normal distribution, mu being the mean vecter while
A being the covariance matrix}

\item{init}{The initial value vector, which indicates the dimensions.}

\item{burn}{Times of iterations one wants to omit before recording.}
}
\value{
A "mcmcn" object `list("chain" = chain)` with chain storing samples by row.
}
\description{
MCMC method for generating vectors with multivariate normal distribution
using Gibbs sampler.
}
\examples{
# Generating Multivariate Normal Distribution Samples------------------------

# provide some parameters
# mu <- c(1,3)
# A <- matrix(c(1, 0.1, 0.1, 1), nrow = 2)

# generating random variates using function `gibbs_norm`
x.norm <- gbs_norm(10000, c(1,3), matrix(c(1, 0.1, 0.1, 1), nrow = 2), c(1,3), burn = 0)

# exploring the results
summary(x.norm)
plot(x.norm)
qqnorm(x.norm$chain[, 1], main = "QQ plot, 1st variable")
}
