% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtrp_exp.R
\name{mtrp_exp}
\alias{mtrp_exp}
\title{Metropolis-Hastings Algorithm Using Exponential Distribution}
\usage{
mtrp_exp(f, n, init, a = 0, rate = 1, burn = 1000)
}
\arguments{
\item{f}{Density function from which one wants to sample.}

\item{n}{The numbers of samples one wants to obtain.}

\item{init}{The initial value vector, which indicates the dimensions.}

\item{a}{The left support boundary vector. if a numeric number is provided,
it will be broadcast to a vector with the same length as init.}

\item{rate}{A vector with rate[i] being the rate of
proposal for updating 'i'th variable. if a numeric number is provided,
it will be broadcast to a vector with the same length as init.}

\item{burn}{Times of iterations one wants to omit before recording.}
}
\value{
A "mcmcn" object `list("chain" = chain, "alpha" = k/iters, "acpt" = acpt)` with
chain storing samples by row, alpha being the rejection rate,
acpt being whether to be accepted each iter.
}
\description{
MCMC method for continuous random vector with exponential distribution as
proposal distribution using an independence sampler algorithm.
}
\examples{
# the Exp(1) Distribution Left-Truncated at 1--------------------------------
(i.e., with support [1,+Infinity))

# pdf f(x) = exp(1 - x)
f <- function(x){
  ifelse(x >= 1, return(exp(-x)), 0)
}

# generating random variates using function `mtrp_exp`
x.exp1 <- mtrp_exp(f, 10000, 1, a = 1, burn = 0)

# exploring the results
cat("The sample rejection rate is", x.exp1$alpha, "\\n")
par(mfrow = c(2, 1))
plot(1:500, x.exp1$chain[1:500], type = "l",
     main = "first 500 iters", ylab = "X")
plot(9501:10000, x.exp1$chain[9501:10000], type = "l",
     main = "last 500 iters", ylab = "X")
par(mfrow = c(1, 1))
hist(x.exp1$chain, freq = FALSE, breaks = 50,
     main = "Histogram of Samples",
     xlab = "X", xlim = c(0,4), ylim = c(0,  exp(1)))
curve(exp(1 - x), col = "red", add = TRUE)
}
