% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdff.R
\name{pdff}
\alias{pdff}
\alias{pdff.norm}
\alias{pdff.mixnorm}
\title{Getting the pdf of some distribution}
\usage{
pdff(type = c("norm", "mixnorm"), ...)

\method{pdff}{norm}(type, mu = 0, A = 1)

\method{pdff}{mixnorm}(type, p, mu1, mu2, A1, A2)
}
\arguments{
\item{type}{Type of the distribution one wants to get.
c("norm", "mixnorm") is provided.}

\item{...}{Additional arguments for type.}

\item{mu, A}{Parameters for normal distribution,
mu being the mean vecter while A being the covariance matrix}

\item{p, m1, mu2, A1, A2}{Parameters for mixture normal distribution.
The pdf can be written as p*pdf1 + (1-p)*pdf2}
}
\value{
The pdf of `type`.
}
\description{
Getting the pdf of some distribution, where the Const is ignored for MCMC method.
}
\examples{
# norm --------------------------------------------------------------------
f <- pdff("norm")
# mixnorm -----------------------------------------------------------------
p <- 0.5
mu1 <- c(0, 0)
mu2 <- c(6, 6)
A1 <- matrix(c(1, 0.1, 0.1, 1), nrow = 2)
A2 <- matrix(c(1, -0.1, -0.1, 1), nrow = 2)
g <- pdff("mixnorm", p, mu1, mu2, A1, A2)
}
