% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbs_norm.R
\name{gbs_norm}
\alias{gbs_norm}
\title{Gibbs Algorithm for Generating Multivariate Normal Distribution}
\usage{
gbs_norm(n, mu, A, init, burn = 1000)
}
\arguments{
\item{n}{The numbers of samples one wants to obtain.}

\item{mu, A}{Parameters for normal distribution, mu being the mean vecter while
A being the covariance matrix}

\item{init}{The initial value vector, which indicates the dimensions.}

\item{burn}{Times of iterations one wants to omit before recording.}
}
\value{
A "mcmcn" object `list("chain" = chain)` with chain storing samples by row.
}
\description{
MCMC method for generating vectors with multivariate normal distribution
using Gibbs sampler.
}
\examples{
# Generating Multivariate Normal Distribution Samples------------------------

# provide some parameters
mu <- c(1,3)
A <- matrix(c(1, 0.1, 0.1, 1), nrow = 2)

# generating random variates using function `gibbs_norm`
x.norm <- gbs_norm(mu, A, 10000, mu, burn = 0)

# exploring the results
cat("The sample mean is", colMeans(x.norm), "\\n")
cat("The sample covariance matrix is\\n")
cov(x.norm)
par(mfrow = c(1, 2))
plot(x.norm[1:100, 1], x.norm[1:100, 2], type = "b", pch = 20,
     main = "first 100 iters", xlab = "1st variable", ylab = "2nd variable")
plot(x.norm[9901:10000, 1], x.norm[9901:10000, 2], type = "b",
     pch = 20, main = "last 100 iters", xlab = "1st variable", ylab = "2nd variable")
par(mfrow = c(2, 1))
plot(1:500, x.norm[1:500, 1], type = "l", ylab = "1st variable",
     main = "first 500 iters")
plot(9501:10000, x.norm[9501:10000, 1], type = "l", ylab = "1st variable",
     main = "last 500 iters")
par(mfrow = c(1, 2))
plot(x.norm, pch = 20, cex = 0.5, main = "Sample Distribution",
     xlab = "1st variable", ylab = "2nd variable")
qqnorm(x.norm[, 1], main = "QQ plot, 1st variable")
par(mfrow = c(1, 1))
}
